+++ lib/src/typedef/typedef.dart
@@ -14,7 +14,8 @@ typedef BeanDecorator<BeanT> = BeanT Function(BeanT);
 
 typedef ListDecorator<BeanT> = List<BeanDecorator<BeanT>>;
 
-typedef BeanInterceptor<BeanT extends Object> = DDIInterceptor<BeanT> Function();
+typedef BeanInterceptor<BeanT extends Object> = DDIInterceptor<BeanT>
+    Function();
 
 typedef ListDDIInterceptor<BeanT extends Object> = List<BeanInterceptor<BeanT>>;