+++ test/beans_test/application_future_test.dart
@@ -12,7 +12,8 @@ import '../clazz_samples/undestroyable/future_application_destroy_get.dart';
 void applicationFuture() {
   group('DDI Application Future Basic Tests', () {
     void registerApplicationBeans() {
-      DDI.instance.registerApplication<A>(() async => A(await DDI.instance.getAsync<B>()));
+      DDI.instance.registerApplication<A>(
+          () async => A(await DDI.instance.getAsync<B>()));
       DDI.instance.registerApplication<B>(() async {
         await Future.delayed(const Duration(milliseconds: 200));
         return B(DDI.instance());
@@ -55,7 +56,8 @@ void applicationFuture() {
       removeApplicationBeans();
     });
 
-    test('Retrieve Application bean after a second "child" bean is diposed', () async {
+    test('Retrieve Application bean after a second "child" bean is diposed',
+        () async {
       registerApplicationBeans();
 
       final instance = await DDI.instance.getAsync<A>();
@@ -70,7 +72,8 @@ void applicationFuture() {
       removeApplicationBeans();
     });
 
-    test('Retrieve Application bean after the last "child" bean is diposed', () async {
+    test('Retrieve Application bean after the last "child" bean is diposed',
+        () async {
       registerApplicationBeans();
 
       final instance1 = await DDI.instance.getAsync<A>();
@@ -142,32 +145,40 @@ void applicationFuture() {
 
       DDI.instance.destroy<C>();
 
-      expect(() => DDI.instance.getAsync<C>(), throwsA(isA<BeanNotFoundException>()));
+      expect(() => DDI.instance.getAsync<C>(),
+          throwsA(isA<BeanNotFoundException>()));
     });
 
     test('Create, get and remove a qualifier bean', () {
-      DDI.instance.registerApplication(() => Future.value(C()), qualifier: 'typeC');
+      DDI.instance
+          .registerApplication(() => Future.value(C()), qualifier: 'typeC');
 
       DDI.instance.getAsync(qualifier: 'typeC');
 
       DDI.instance.destroy(qualifier: 'typeC');
 
-      expect(() => DDI.instance.getAsync(qualifier: 'typeC'), throwsA(isA<BeanNotFoundException>()));
+      expect(() => DDI.instance.getAsync(qualifier: 'typeC'),
+          throwsA(isA<BeanNotFoundException>()));
     });
 
     test('Try to destroy a undestroyable Application bean', () async {
-      DDI.instance.registerApplication(() => Future.value(FutureApplicationDestroyGet()), destroyable: false);
+      DDI.instance.registerApplication(
+          () => Future.value(FutureApplicationDestroyGet()),
+          destroyable: false);
 
-      final instance1 = await DDI.instance.getAsync<FutureApplicationDestroyGet>();
+      final instance1 =
+          await DDI.instance.getAsync<FutureApplicationDestroyGet>();
 
       DDI.instance.destroy<FutureApplicationDestroyGet>();
 
-      final instance2 = await DDI.instance.getAsync<FutureApplicationDestroyGet>();
+      final instance2 =
+          await DDI.instance.getAsync<FutureApplicationDestroyGet>();
 
       expect(instance1, same(instance2));
     });
     test('Register and retrieve Future Application', () async {
-      DDI.instance.registerApplication(() async => A(await DDI.instance.getAsync<B>()));
+      DDI.instance
+          .registerApplication(() async => A(await DDI.instance.getAsync<B>()));
       DDI.instance.registerApplication(() async => B(DDI.instance()));
       DDI.instance.registerApplication(C.new);
 
@@ -203,7 +214,8 @@ void applicationFuture() {
       DDI.instance.registerApplication(C.new);
 
       //This happens because A(await DDI.instance()) transform to A(await DDI.instance<FutureOr<B>>())
-      expect(() => DDI.instance.getAsync<A>(), throwsA(isA<BeanNotFoundException>()));
+      expect(() => DDI.instance.getAsync<A>(),
+          throwsA(isA<BeanNotFoundException>()));
 
       DDI.instance.destroy<A>();
       DDI.instance.destroy<B>();
@@ -211,7 +223,8 @@ void applicationFuture() {
     });
 
     test('Register and retrieve Application bean using FutureOr', () async {
-      DDI.instance.registerApplication(() async => A(await DDI.instance.getAsync()));
+      DDI.instance
+          .registerApplication(() async => A(await DDI.instance.getAsync()));
       DDI.instance.registerApplication<B>(() async => B(DDI.instance()));
       DDI.instance.registerApplication(C.new);
 
@@ -227,8 +240,11 @@ void applicationFuture() {
       DDI.instance.destroy<B>();
       DDI.instance.destroy<C>();
     });
-    test('Retrieve Application bean after a "child" bean is disposed using Future', () async {
-      DDI.instance.registerApplication(() async => A(await DDI.instance.getAsync<B>()));
+    test(
+        'Retrieve Application bean after a "child" bean is disposed using Future',
+        () async {
+      DDI.instance
+          .registerApplication(() async => A(await DDI.instance.getAsync<B>()));
       DDI.instance.registerApplication<B>(() async => B(DDI.instance()));
       DDI.instance.registerApplication(C.new);