+++ test/beans_test/dispose_destroy_all_session_test.dart
@@ -23,9 +23,12 @@ void disposeDestroyAllSession() {
 
       DDI.instance.destroyAllSession();
 
-      expect(() => DDI.instance.get<A>(), throwsA(isA<BeanNotFoundException>()));
-      expect(() => DDI.instance.get<B>(), throwsA(isA<BeanNotFoundException>()));
-      expect(() => DDI.instance.get<C>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<A>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<B>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<C>(), throwsA(isA<BeanNotFoundException>()));
     });
 
     test('Register and retrieve Factory Session bean', () {
@@ -37,9 +40,12 @@ void disposeDestroyAllSession() {
 
       DDI.instance.destroyAllSession();
 
-      expect(() => DDI.instance.get<A>(), throwsA(isA<BeanNotFoundException>()));
-      expect(() => DDI.instance.get<B>(), throwsA(isA<BeanNotFoundException>()));
-      expect(() => DDI.instance.get<C>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<A>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<B>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<C>(), throwsA(isA<BeanNotFoundException>()));
     });
 
     test('Register, get, dispose and destroy Session bean', () {
@@ -64,13 +70,17 @@ void disposeDestroyAllSession() {
 
       DDI.instance.destroyAllSession();
 
-      expect(() => DDI.instance.get<A>(), throwsA(isA<BeanNotFoundException>()));
-      expect(() => DDI.instance.get<B>(), throwsA(isA<BeanNotFoundException>()));
-      expect(() => DDI.instance.get<C>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<A>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<B>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<C>(), throwsA(isA<BeanNotFoundException>()));
     });
 
     test('Register, get, dispose and destroy Session bean', () async {
-      DDI.instance.registerSession(() async => A(await DDI.instance.getAsync()));
+      DDI.instance
+          .registerSession(() async => A(await DDI.instance.getAsync()));
       DDI.instance.registerSession<B>(() => Future.value(B(DDI.instance())));
       DDI.instance.registerSession(C.new);
 
@@ -93,9 +103,12 @@ void disposeDestroyAllSession() {
 
       DDI.instance.destroyAllSession();
 
-      expect(() async => DDI.instance.getAsync<A>(), throwsA(isA<BeanNotFoundException>()));
-      expect(() async => DDI.instance.getAsync<B>(), throwsA(isA<BeanNotFoundException>()));
-      expect(() async => DDI.instance.getAsync<C>(), throwsA(isA<BeanNotFoundException>()));
+      expect(() async => DDI.instance.getAsync<A>(),
+          throwsA(isA<BeanNotFoundException>()));
+      expect(() async => DDI.instance.getAsync<B>(),
+          throwsA(isA<BeanNotFoundException>()));
+      expect(() async => DDI.instance.getAsync<C>(),
+          throwsA(isA<BeanNotFoundException>()));
     });
   });
 }