+++ test/beans_test/factory_interceptor_test.dart
@@ -65,7 +65,8 @@ void factoryInterceptor() {
 
       ddi.destroy<G>(qualifier: 'qualifier');
 
-      expect(() => ddi.get<G>(qualifier: 'qualifier'), throwsA(isA<BeanNotFoundException>()));
+      expect(() => ddi.get<G>(qualifier: 'qualifier'),
+          throwsA(isA<BeanNotFoundException>()));
     });
 
     test('ADD Interceptor to a Factory Dependent bean', () {