+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Warnings:
     * When registering a Factory Future and trying to obtain more than one instance simultaneously, it may cause a race condition and will be blocked. Especially if the instance is Dependent Scope.
+    * When using Interceptors and Factories, you must register your Factory with `ddi.register(factoryClazz: FactoryClazz.application(clazzFactory: ..., interceptors: [...]))`
 
 * Break changes:
     * `DDIInterceptor.aroundConstruct` renamed to `DDIInterceptor.onCreate`.